backend-supply-chain/requirements.txt
fastapi
uvicorn
mysql-connector-python
SQLAlchemy
Authlib
python-dotenv
PyJWT
cryptography 
